+++ assignments/2025-03-13/src/solution1/README.md
@@ -9,8 +9,6 @@ Cleric 클래스에 “이름", “HP”, “최대 HP”, “MP”, “최대 M
 또한 HP와 최대 HP는 정수로 초기치 50, MP와 최대 MP는 정수로 초기치 10으로 하고,
 
 최대 HP와 최대 MP는 상수 필드로 선언 하시오.
-
-
 ---
 ### **1. 기능 개요**
 
@@ -40,3 +38,41 @@ Cleric 클래스에 “이름", “HP”, “최대 HP”, “MP”, “최대 M
 - `maxHp`와 `maxMp`는 생성자에서 변경할 수 없다.
 - `hp`와 `mp`가 음수가 되는 상황을 방지해야 한다. (나중에 데미지를 입었을 때)
 - 생성자를 통해 `name`을 필수 입력값으로 설정한다.
+
+
+---
+# 연습문제 2
+
+성직자는 “셀프 에이드" 라는 마법을 사용할 수 있고, MP를 5소비하는 것으로 자신의 HP 를 최대 HP 까지 회복할 수 있다.
+
+연습 1 에 선언한 Cleric 클래스에 “selfAid()” 메소드를 추가 하시오.
+
+또한, 이 메소드는 인수가 없고, 리턴 값도 없다.
+---
+
+### **1. 기능 개요**
+
+- 성직자는 **"셀프 에이드(Self Aid)"** 라는 마법을 사용할 수 있다.
+- **MP를 5 소비**하고, **HP를 최대 HP(`maxHp`)까지 회복**한다.
+- `selfAid()` 메서드는 **매개변수 없음**, **반환값 없음**.
+
+### **2. `selfAid()` 메서드의 동작 조건**
+
+- **MP가 5 이상**이면:
+    - MP를 5 소모하고, HP를 `maxHp`로 회복.
+- **MP가 5 미만**이면:
+    - 아무 동작도 하지 않음.
+
+### **3. 변경되는 속성**
+
+| 속성명 | 변경 여부 | 변경 조건 |
+| --- | --- | --- |
+| `hp` | **변경됨** | `hp = maxHp` (MP가 충분할 때) |
+| `mp` | **변경됨** | `mp -= 5` (MP가 충분할 때) |
+
+### **4. 요구사항 및 제약사항**
+
+- `selfAid()` 메서드는 **외부 입력 없이 호출만 하면 실행**되어야 함.
+- `hp`가 `maxHp`보다 커질 수 없음.
+- MP가 부족할 경우 아무 동작도 하지 않음.
+---
